reorganize components, capitalize component file names
