+++ escuela/src/components/addClass/AddClassPage.js
