+++ escuela/src/components/landingPage/CardImg.js
