+++ escuela/src/components/landingPage/Enroll.js
