+++ escuela/src/components/landingPage/Events.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Row, Container } from "reactstrap";
-import CardImg from "./cardImg";
+import CardImg from "./CardImg";
 import LandingCard from "./LandingCard";
 
 export const Events = () => {
@@ -17,13 +17,13 @@ export const Events = () => {
           btnText="Got to Events"
           buttonClass="btn-md"
         />
-    
+
         <CardImg
           src="https://images.unsplash.com/photo-1523958399470-c15bf82fcea9?ixlib=rb-0.3.5&ixid=eyJhcHBfaWQiOjEyMDd9&s=24bc33ab52ef2dc36a177c25c24da122&auto=format&fit=crop&w=800&q=60"
           alt="events"
           className="col-5 h-25 ml-5"
         />
-     
+
       </div>
     </Container>
   );
