+++ escuela/src/components/landingPage/LandingCard.js
