+++ escuela/src/components/landingPage/SliderForm.js
